unimportant

+++ shlib/fs.sh
@@ -1,26 +1,67 @@
 #use die
 
-rm_dir() { # dir
-	[ "${1:0:1}" == "/" ] || { say "path not absolute"; return 1; }
-	[ "${1: -1}" == "/" ] && { say "path ends in slash"; return 1; }
-	local dir="$1"
-	say "Removing dir $dir ..."
-	must rm -rf "$dir"
+check_abs_filepath() {
+	[ "${1:0:1}" == "/" ] || die "path not absolute: $1"
+	[ "${1: -1}" == "/" ] && die "path ends in slash: $1"
 }
 
-cp_dir() { # src_dir dst_dir
+rm_dir() { # DIR
+	local dir="$1"
+	checkvars dir
+	check_abs_filepath "$dir"
+	say -n "Removing dir $dir ... "
+	[ "$DRY" ] || must rm -rf "$dir"
+	say OK
+}
+
+cp_dir() { # SRC_DIR DST_DIR
 	local src_dir="$1"
 	local dst_dir="$2"
 	checkvars src_dir dst_dir
-	[ "${src_dir:0:1}" == "/" ] || { say "src dir not absolute"; return 1; }
-	say "Removing dir $dst_dir ..."
-	must rm -rf "$dst_dir"
-	say "Copying dir $src_dir to $dst_dir ..."
-	must cp -rf "$src_dir" "$dst_dir"
+	check_abs_filepath "$src_dir"
+	check_abs_filepath "$dst_dir"
+	[ -d "$dst_dir" ] && rm_dir "$dst_dir"
+	say -n "Copying dir $src_dir to $dst_dir ... "
+	[ "$DRY" ] || must cp -rf "$src_dir" "$dst_dir"
+	say OK
 }
 
-sha_dir() { # dir
+sha_dir() { # DIR
 	local dir="$1"
 	checkvars dir
 	find "$dir" -type f -print0 | LC_ALL=C sort -z | xargs -0 sha1sum | sha1sum | cut -d' ' -f1
+}
+
+append() { # S FILE
+	local s="$1"
+	local file="$2"
+	checkvars s- file
+	say -n "Appending ${#s} bytes to file $file ... "
+	debug -n "MUST: append \"$s\" $file "
+	if [ "$DRY" ] || printf "%s" "$s" >> "$file"; then
+		debug "[$?]"
+	else
+		die "append $file [$?]"
+	fi
+	say OK
+}
+
+save() { # S FILE [USER]
+	local s="$1"
+	local file="$2"
+	local user="$3"
+	checkvars s- file
+	say -n "Saving ${#s} bytes to file $file ... "
+	debug -n "MUST: save \"$s\" $file "
+	if [ "$DRY" ] || printf "%s" "$s" > "$file"; then
+		debug "[$?]"
+	else
+		die "save $file [$?]"
+	fi
+	if [ "$user" ]; then
+		checkvars user
+		must chown $user:$user $file
+		must chmod 600 $file
+	fi
+	say OK
 }
